backend/requirements.txt
pymongo>=4.6.0
dnspython>=2.4.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
scikit-learn>=1.6.1
xgboost>=2.0.0
joblib>=1.3.0
numpy>=1.24.0
atproto>=0.0.30
geopy>=2.3.0
